pyproject.toml
[tool.poetry]
name = "toconvert"
version = "0.1.0"
description = ""
authors = ["Michel Kreutner <mkreutner@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai-whisper = "^20231117"
setuptools-rust = "^1.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
